requirements.txt

# Core dependencies for Framepack Generator Pro
gradio==4.44.0
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

# Image processing and computer vision
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# AI/ML libraries
transformers==4.35.2
sentence-transformers==2.2.2
scikit-learn==1.3.2

# API clients for multi-provider support
openai>=1.0.0
google-generativeai>=0.8.0

# Data handling
pandas==2.1.3
pydantic==2.5.0

# HTTP requests and utilities
requests==2.31.0
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
colorama==0.4.6

# Development and testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Additional utilities
python-multipart>=0.0.9
jinja2==3.1.2
